+++ Gogaman/src/Gogaman/Rendering/Texture/AbstractTexture3D.h
@@ -15,7 +15,7 @@ namespace Gogaman
 		AbstractTexture3D &operator=(const AbstractTexture3D &) = delete;
 		AbstractTexture3D &operator=(AbstractTexture3D &&other) = default;
 
-		void Generate(const int width, const int height, const int depth, const uint8_t *imageData = nullptr) { this->GetImplementation().Generate(width, height, depth, imageData); }
+		void Generate(const uint16_t width, const uint16_t height, const uint16_t depth, const uint8_t *imageData = nullptr) { this->GetImplementation().Generate(width, height, depth, imageData); }
 	protected:
 		AbstractTexture3D()
 			: width(0), height(0), depth(0), wrapS(TextureWrapMode::Repeat), wrapT(TextureWrapMode::Repeat), wrapR(TextureWrapMode::Repeat)
@@ -23,7 +23,7 @@ namespace Gogaman
 
 		~AbstractTexture3D() = default;
 	public:
-		int             width, height, depth;
+		uint16_t        width, height, depth;
 		TextureWrapMode wrapS, wrapT, wrapR;
 	};
 }
